server/toota/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.3
daphne==4.1.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.7
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-django==4.8.0
pytz==2024.1
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
zope.interface==6.2
